--- src/database/DAL/__init__.py
@@ -1,6 +1,0 @@
-# SPDX-License-Identifier: MIT
-# SPDX-FileCopyrightText: 2023 Sophie Heasman <sophieheasmann@gmail.com>
-
-from .database_abstraction_layer import DataAbstractionLayer
-from .local import LocalDatabase
-from .s3 import S3Database
