--- src/database/DAL/local.py
@@ -1,72 +1,0 @@
-# SPDX-License-Identifier: MIT
-# SPDX-FileCopyrightText: 2023 Sophie Heasman <sophieheasmann@gmail.com>
-
-import json
-import os
-
-import pandas as pd
-
-from database.DAL import DataAbstractionLayer
-from logger import get_logger
-
-log = get_logger()
-
-
-class LocalDatabase(DataAbstractionLayer):
-    DF_INPUT = "./data/sumup_leads_email.csv"
-    DF_OUTPUT = "./data/leads_enriched.csv"
-    REVIEWS = "./data/reviews/"
-
-    def _download(self):
-        """
-        Download database from specified DF path
-        """
-        try:
-            self.df = pd.read_csv(self.DF_INPUT)
-        except FileNotFoundError:
-            log.error("Error: Could not find input file for Pipeline.")
-
-    def save_dataframe(self):
-        """
-        Save dataframe in df attribute in chosen output location
-        """
-        self.df.to_csv(self.DF_OUTPUT, index=False)
-        log.info(f"Saved enriched data locally to {self.DF_OUTPUT}")
-
-    def insert_data(self, data):
-        """
-        TODO: Insert new data into specified dataframe
-        :param data: Data to be inserted (desired format must be checked)
-        """
-        pass
-
-    def save_review(self, review, place_id, force_refresh=False):
-        """
-        Upload review to specified review path
-        :param review: json contents of the review to be uploaded
-        """
-        # Write the data to a JSON file
-        file_name = place_id + "_reviews.json"
-        json_file_path = self.REVIEWS + file_name
-
-        if os.path.exists(json_file_path):
-            log.info(f"Reviews for {place_id} already exist")
-            return
-
-        with open(json_file_path, "w", encoding="utf-8") as json_file:
-            json.dump(review, json_file, ensure_ascii=False, indent=4)
-
-    def fetch_review(self, place_id):
-        """
-        Fetch review for specified place_id
-        :return: json contents of desired review
-        """
-        reviews_path = self.REVIEWS + place_id + "_reviews.json"
-        try:
-            with open(reviews_path, "r", encoding="utf-8") as reviews_json:
-                reviews = json.load(reviews_json)
-                return reviews
-        except:
-            log.warning(f"Error loading reviews from path {reviews_path}.")
-            # Return empty list if any exception occurred or status is not OK
-            return []
