+++ src/database/__init__.py
@@ -2,7 +2,7 @@
 # SPDX-FileCopyrightText: 2023 Felix Zailskas <felixzailskas@gmail.com>
 
 from config import DATABASE_TYPE
-from database.DAL import DataAbstractionLayer, LocalDatabase, S3Database
+from database.leads import LocalRepository, Repository, S3Repository
 from logger import get_logger
 
 from .database_dummy import DatabaseDummy
@@ -12,13 +12,13 @@ _database = None
 log = get_logger()
 
 
-def get_database() -> DataAbstractionLayer:
+def get_database() -> Repository:
     global _database
     if _database is None:
         if DATABASE_TYPE == "S3":
-            _database = S3Database()
+            _database = S3Repository()
         elif DATABASE_TYPE == "Local":
-            _database = LocalDatabase()
+            _database = LocalRepository()
         else:
             log.error("Database type not initialised")
             raise ValueError
